backend/requirements.txt
absl-py==2.1.0
addict==2.4.0
altair==5.2.0
annotated-types==0.6.0
anyio==4.1.0
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.1.0
Babel==2.13.1
beautifulsoup4==4.12.2
bleach==6.1.0
blinker==1.7.0
cachetools==5.3.2
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==3.1.0
click==8.1.7
comm==0.2.0
contourpy==1.0.5
cvxopt==1.3.2
cycler==0.11.0
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.19.0
filelock==3.16.1
fonttools==4.37.3
fpdf==1.7.2
fpdf2==2.7.9
fqdn==1.5.1
frozendict==2.4.4
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.40
graphviz==0.20.3
grpcio==1.66.2
html5lib==1.1
idna==3.4
importlib-metadata==6.11.0
inflection==0.5.1
ics==0.7.2
iniconfig==1.1.1
intervaltree==3.1.0
ipykernel==6.27.1
ipytest==0.14.2
ipython==8.18.1
irreceiver==0.9.5
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyter_server==2.11.1
jupyter_server_terminals==0.4.4
jupyterlab==4.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
kiwisolver==1.4.4
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.9.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
mpmath==1.3.0
multitasking==0.0.11
nbclient==0.7.4
nbconvert==7.11.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==2.8.8
notebook==7.0.6
notebook_shim==0.2.3
numpy==2.1.3
onnx==1.17.0
opencv-python==4.6.0.66
openpyxl==3.1.5
overrides==7.4.0
packaging==21.3
paho-mqtt==1.6.1
pandas==2.2.3
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.4
peewee==3.17.6
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.0.0
plotly==5.24.1
pluggy==1.0.0
pm4py==2.7.11.13
prometheus-client==0.19.0
prompt-toolkit==3.0.41
protobuf==4.25.1
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
py-cpuinfo==9.0.0
pyairtable==2.2.0
pyarrow==14.0.1
pycocotools==2.0.8
pycparser==2.21
pycryptodome==3.17
pydantic==2.5.2
pydantic_core==2.14.5
pydeck==0.8.1b0
pydotplus==2.0.2
Pygments==2.17.2
pyparsing==3.0.9
pytest==7.1.3
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.7
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
referencing==0.31.0
reportlab==4.0.7
repoze.lru==0.7
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
Routes==2.5.1
rpds-py==0.13.1
scipy==1.11.4
seaborn==0.13.0
Send2Trash==1.8.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
stack-data==0.6.3
statsmodels==0.14.0
streamlit==1.30.0
sympy==1.13.3
tenacity==8.2.3
tensorboard==2.18.0
tensorboard-data-server==0.7.2
terminado==0.18.0
thop==0.1.1.post2209072238
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.4.1
torchvision==0.19.1
tornado==6.3.3
tqdm==4.66.5
traitlets==5.14.0
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
ultralytics==8.3.7
ultralytics-thop==2.0.9
uri-template==1.3.0
urllib3==1.26.15
validators==0.22.0
voila==0.5.5
watchdog==3.0.0
wcwidth==0.2.12
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
websockets==12.0
Werkzeug==3.0.4
yfinance==0.2.41
zipp==3.17.0
